+++ tests/Moq.Tests/ProtectedAsMockFixture.cs
@@ -131,6 +131,13 @@ namespace Moq.Tests
 		}
 
 		[Fact]
+		public void Setup_can_automock()
+		{
+			this.protectedMock.Setup(m => m.Nested.Method(1)).Returns(123);
+			Assert.Equal(123, mock.Object.GetNested().Method(1));
+		}
+
+		[Fact]
 		public void SetupGet_can_setup_readonly_property()
 		{
 			this.protectedMock.SetupGet(m => m.ReadOnlyPropertyImpl).Returns(42);
@@ -146,6 +153,25 @@ namespace Moq.Tests
 			this.protectedMock.SetupGet(m => m.ReadWritePropertyImpl).Returns(42);
 
 			var actual = this.mock.Object.ReadWriteProperty;
+
+			Assert.Equal(42, actual);
+		}
+
+		[Fact]
+		public void SetUpGet_can_automock()
+		{
+			this.protectedMock.SetupGet(m => m.Nested.Value).Returns(42);
+
+			var actual = mock.Object.GetNested().Value;
+
+			Assert.Equal(42, actual);
+		}
+
+		[Fact] void SetupGet_can_setup_virtual_property()
+		{
+			this.protectedMock.SetupGet(m => m.Virtual).Returns(42);
+
+			var actual = mock.Object.GetVirtual();
 
 			Assert.Equal(42, actual);
 		}
@@ -298,6 +324,12 @@ namespace Moq.Tests
 			Assert.Contains("Was not queried.", exception.Message);
 		}
 
+		public interface INested
+		{
+			int Value { get; }
+			int Method(int value);
+		}
+
 		public abstract class Foo
 		{
 			protected Foo()
@@ -336,6 +368,32 @@ namespace Moq.Tests
 			protected abstract void DoSomethingImpl(int arg);
 
 			protected abstract int GetSomethingImpl();
+
+			protected abstract INested Nested { get; set; }
+
+			public INested GetNested()
+			{
+				return Nested;
+			}
+
+			private int virtualProperty;
+			public virtual int Virtual
+			{
+				protected get
+				{
+					return virtualProperty;
+				}
+				set
+				{
+					virtualProperty = value;
+				}
+
+			}
+
+			public int GetVirtual()
+			{
+				return Virtual;
+			}
 		}
 
 		public interface Fooish
@@ -347,6 +405,8 @@ namespace Moq.Tests
 			void DoSomethingImpl(int arg);
 			int GetSomethingImpl();
 			void NonExistentMethod();
+			INested Nested { get; set; }
+			int Virtual { get; set; }
 		}
 
 		public abstract class MessageHandlerBase